Feature/update delete pantry

+++ app/(site)/pantry/[id]/actions.ts
@@ -3,7 +3,13 @@
 import { prisma } from "@/lib/prisma";
 import { getCurrentUser } from "@/lib/session";
 
-export async function getPantry({ pantryId }: { pantryId: string }) {
+export async function getPantry({
+  pantryId,
+  needListItems,
+}: {
+  pantryId: string;
+  needListItems: boolean;
+}) {
   const user = await getCurrentUser();
 
   if (!user) {
@@ -17,7 +23,8 @@ export async function getPantry({ pantryId }: { pantryId: string }) {
     },
     select: {
       name: true,
-      listItems: true,
+      description: true,
+      listItems: needListItems,
     },
   });
 
@@ -42,7 +49,15 @@ export async function getPantryItems({ pantryId }: { pantryId: string }) {
   return data;
 }
 
-export async function deletePantry({ pantryId }: { pantryId: string }) {
+export async function updatePantry({
+  pantryId,
+  pantryName,
+  pantryDescription,
+}: {
+  pantryId: string;
+  pantryName: string;
+  pantryDescription: string;
+}) {
   const user = await getCurrentUser();
 
   if (!user) {
@@ -50,11 +65,36 @@ export async function deletePantry({ pantryId }: { pantryId: string }) {
     return;
   }
 
-  const data = await prisma.list.delete({
+  const data = await prisma.list.update({
+    data: {
+      name: pantryName,
+      description: pantryDescription,
+    },
     where: {
       id: pantryId,
     },
   });
 
   return data;
+}
+
+export async function deletePantry({ pantryId }: { pantryId: string }) {
+  try {
+    const user = await getCurrentUser();
+
+    if (!user) {
+      console.log("unauthorized");
+      return;
+    }
+
+    const data = await prisma.list.delete({
+      where: {
+        id: pantryId,
+      },
+    });
+
+    return data;
+  } catch (error) {
+    console.log("error in deletePantry action: ", error);
+  }
 }
